requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap5==2024.10
Django==5.0.9
django-allauth==65.2.0
django-ckeditor==6.7.1
django-crispy-forms==2.3
django-jalali-date==1.1.3
django-js-asset==2.2.0
django-rosetta==0.10.1
environs==11.2.1
idna==3.10
jalali_core==1.0.0
jdatetime==5.0.0
marshmallow==3.23.1
mssql-django==1.5
mysqlclient==2.2.6
packaging==24.2
pillow==11.0.0
polib==1.2.0
pyodbc==5.2.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
setuptools==75.5.0
sqlparse==0.5.1
urllib3==2.2.3
